% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasheet.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Convert a Matrix to a Tibble}
\usage{
matrix_to_df(m)
}
\arguments{
\item{m}{A character matrix where the first row contains column headers.}
}
\value{
A tibble representing the data in the matrix.
}
\description{
Converts a character matrix to a tibble. The first row is treated as column headers.
Columns with empty names are removed, and empty values are replaced with \code{NA}.
}
