% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_datasheet.R
\name{parse_sheet}
\alias{parse_sheet}
\title{Parse a data sheet}
\usage{
parse_sheet(sheet)
}
\arguments{
\item{sheet}{A character vector in which each element is a (tab-delimited) row from a data sheet}
}
\value{
A named list of \code{tibble} objects or an error message if the parser fails
}
\description{
Parses tables from a data sheet and returns a list with named tables.
}
